Reorganizado local de get items da api
